refactor(tests): split executor_test.go into focused modules
Remove internal/executor/executor_test.go from the file-size grandfathered
list and split its 1 732 lines into five focused files, each under 500 lines:

- executor_helpers_test.go  — shared test infrastructure (bufconn, mockAgentServer,
                              testEnv, setupTestEnv, registerHealthyAgent)
- executor_test.go          — basic execution paths (success, agent status, config
                              propagation, context passing, unexpected status)
- executor_options_test.go  — constructor defaults, WithTimeout / WithMaxRetries /
                              WithDeadlineMode option clamping
- executor_retry_test.go    — isTransient table, retry/exhaustion/concurrency tests,
                              context-cancel mid-dispatch, dial-options additive check
- executor_deadline_test.go — derived-deadline computation, shared retry budget,
                              minimum budget cutoff, static-mode preservation
- executor_metadata_test.go — wall-time accuracy, retry count, FAILED metadata
                              preservation, int32 overflow guard, token budget cutoff
