+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **421** tracked files
+> **426** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 108 |
+| `internal/` | 113 |
 | `docs/` | 81 |
 | `agents/` | 77 |
 | `tests/` | 65 |
@@ -27,7 +27,7 @@
 | Extension | Count |
 |-----------|-------|
 | `.py` | 153 |
-| `.go` | 105 |
+| `.go` | 110 |
 | `.md` | 97 |
 | `.yaml` | 25 |
 | `.rs` | 9 |
@@ -323,6 +323,11 @@
 │   │   ├── chat_test.go
 │   │   ├── dispatch.go
 │   │   ├── executor.go
+│   │   ├── executor_deadline_test.go
+│   │   ├── executor_helpers_test.go
+│   │   ├── executor_metadata_test.go
+│   │   ├── executor_options_test.go
+│   │   ├── executor_retry_test.go
 │   │   └── executor_test.go
 │   ├── generated/
 │   │   ├── logpb/
